fix test_config text_axon text_keypair text_wallet
