+++ requirements/prod.txt
@@ -15,6 +15,7 @@ pyyaml
 password_strength
 pydantic!=1.8,!=1.8.1,<2.0.0,>=1.7.4
 python-Levenshtein
+pytest
 retry
 requests
 rich
