+++ tests/unit_tests/test_config.py
@@ -15,22 +15,12 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 # DEALINGS IN THE SOFTWARE.
 
-from multiprocessing.sharedctypes import Value
-from sys import prefix
 import argparse
-import pytest
-from unittest.mock import MagicMock
-
 import bittensor
-
-
-def test_loaded_config():
-    with pytest.raises(NotImplementedError):
-        bittensor.Config(loaded_config=True)
+from unittest.mock import MagicMock
 
 def test_strict():
     parser = argparse.ArgumentParser()
-
     # Positional/mandatory arguments don't play nice with multiprocessing.
     # When the CLI is used, the argument is just the 0th element or the filepath.
     # However with multiprocessing this function call actually comes from a subprocess, and so there
@@ -39,14 +29,10 @@ def test_strict():
     parser.add_argument("--cov", help="Dummy Args")
     parser.add_argument("--cov-append", action='store_true', help="Dummy Args")
     parser.add_argument("--cov-config",  help="Dummy Args")
-    #bittensor.Dendrite.add_args( parser )
     bittensor.logging.add_args( parser )
     bittensor.wallet.add_args( parser )
     bittensor.subtensor.add_args( parser )
-    #bittensor.metagraph.add_args( parser )
-    bittensor.dataset.add_args( parser )
     bittensor.axon.add_args( parser )
-    #bittensor.wandb.add_args( parser )
     bittensor.config( parser, strict=False)
     bittensor.config( parser, strict=True)
 
@@ -55,7 +41,7 @@ def test_prefix():
     parser = argparse.ArgumentParser()
 
     mock_wallet = MagicMock(
-        spec=bittensor.Wallet,
+        spec=bittensor.wallet,
         coldkey=MagicMock(),
         coldkeypub=MagicMock(
             # mock ss58 address
@@ -66,9 +52,6 @@ def test_prefix():
         ),
     )
 
-    #bittensor.dendrite.add_args( parser)
-    #bittensor.dendrite.add_args( parser, prefix = 'second' )
-
     bittensor.logging.add_args( parser )
     bittensor.logging.add_args( parser, prefix = 'second' )
 
@@ -78,35 +61,16 @@ def test_prefix():
     bittensor.subtensor.add_args( parser )
     bittensor.subtensor.add_args( parser, prefix = 'second'  )
 
-    #bittensor.metagraph.add_args( parser )
-    #bittensor.metagraph.add_args( parser, prefix = 'second' )
-
-    bittensor.dataset.add_args( parser )
-    bittensor.dataset.add_args( parser, prefix = 'second' )
-
     bittensor.axon.add_args( parser )
     bittensor.axon.add_args( parser, prefix = 'second' )
-
-    #bittensor.wandb.add_args( parser )
-    #bittensor.wandb.add_args( parser, prefix = 'second' )
 
     config_non_strict = bittensor.config( parser, strict=False)
     config_strict = bittensor.config( parser, strict=True)
 
-    #bittensor.dendrite( config_strict ).__del__()
-    #bittensor.dendrite( config_non_strict ).__del__()
-    #bittensor.dendrite( config_strict.second ).__del__()
-    #bittensor.dendrite( config_non_strict.second ).__del__()
-
-    bittensor.axon( metagraph=None, wallet=mock_wallet, config=config_strict ).stop()
-    bittensor.axon( metagraph=None, wallet=mock_wallet, config=config_non_strict ).stop()
-    bittensor.axon( metagraph=None, wallet=mock_wallet, config=config_strict.second ).stop()
-    bittensor.axon( metagraph=None, wallet=mock_wallet, config=config_non_strict.second ).stop()
-
-    #bittensor.metagraph( config_strict )
-    #bittensor.metagraph( config_non_strict )
-    #bittensor.metagraph( config_strict.second )
-    #bittensor.metagraph( config_non_strict.second )
+    bittensor.axon( wallet=mock_wallet, config=config_strict ).stop()
+    bittensor.axon( wallet=mock_wallet, config=config_non_strict ).stop()
+    bittensor.axon( wallet=mock_wallet, config=config_strict.second ).stop()
+    bittensor.axon( wallet=mock_wallet, config=config_non_strict.second ).stop()
 
     bittensor.wallet( config_strict )
     bittensor.wallet( config_non_strict )
@@ -118,33 +82,6 @@ def test_prefix():
     bittensor.logging( config_strict.second )
     bittensor.logging( config_non_strict.second )
 
-    # This is the only place we call bittensor.wandb() outside of neuron code.
-    # It fails because we don't have a key set up for this.
-    # TODO: Actually test bittensor.wandb
-    #bittensor.wandb( config_strict )
-    #bittensor.wandb( config_non_strict )
-    #bittensor.wandb( config_strict.second )
-    #bittensor.wandb( config_non_strict.second )
-
-
-def construct_config():
-    defaults = bittensor.Config()
-    bittensor.subtensor.add_defaults( defaults )
-    #bittensor.dendrite.add_defaults( defaults )
-    bittensor.axon.add_defaults( defaults )
-    bittensor.wallet.add_defaults( defaults )
-    bittensor.dataset.add_defaults( defaults )
-    bittensor.logging.add_defaults( defaults )
-    #bittensor.wandb.add_defaults( defaults )
-
-    return defaults
-
-def test_to_defaults():
-    config = construct_config()
-    config.to_defaults()
 
 if __name__  == "__main__":
-    # test_loaded_config()
-    # test_strict()
-    # test_to_defaults()
     test_prefix()
